Change: replicated the tilemap canvas and added to the 'ProbeBuilder' scene to evaluate how to integrate.
Edited the scale of the new probe items/components and adjusted the camera position.  This step is for visualizing and learning how to work with layers.

Note the spawned probe needed layer adjusted to appear on top of the tilemap.  Calculation of mouse drag/move positioning and collision/connecting the parts may be more complicated with the isometric view--TBD.